requirements.txt
requests
beautifulsoup4
pandas
openpyxl
python-dotenv
openai
anthropic
google-generativeai
googlesearch-python
fastapi
uvicorn
jinja2
secure-smtplib
tenacity
fake-useragent
streamlit
psycopg2-binary
sqlalchemy
cryptography
